rewtire methods and tests
