+++ src/main/java/com/qaprosoft/carina/demo/web/enums/Devices.java
@@ -1,26 +1,16 @@
 package com.qaprosoft.carina.demo.web.enums;
 
-import com.qaprosoft.carina.core.gui.AbstractPage;
-import com.qaprosoft.carina.demo.web.gui.common.PhonePageBase;
-import com.qaprosoft.carina.demo.web.gui.common.TabletsPageBase;
-
 public enum Devices {
-    TABLETS("Планшети",TabletsPageBase.class),
-    PHONES("Мобільні телефони", PhonePageBase.class);
+    TABLETS("Планшети"),
+    PHONES("Мобільні телефони");
 
     private String namePage;
-    private Class<? extends AbstractPage> pageClass;
 
-    Devices(String namePage, Class<? extends AbstractPage> pageClass) {
+    Devices(String namePage) {
         this.namePage = namePage;
-        this.pageClass = pageClass;
     }
 
     public String getNamePage() {
         return namePage;
-    }
-
-    public Class<? extends AbstractPage> getPageClass() {
-        return this.pageClass;
     }
 }
