+++ src/main/java/com/qaprosoft/carina/demo/web/gui/common/AppleBrandPageBase.java
@@ -1,9 +1,9 @@
 package com.qaprosoft.carina.demo.web.gui.common;
 
-import com.qaprosoft.carina.demo.web.enums.AppleDevices;
 import org.openqa.selenium.WebDriver;
 
 import com.qaprosoft.carina.core.gui.AbstractPage;
+import com.qaprosoft.carina.demo.web.enums.AppleDevices;
 import com.qaprosoft.carina.demo.web.gui.components.FooterMenu;
 
 public abstract class AppleBrandPageBase extends AbstractPage {
