+++ src/main/java/com/qaprosoft/carina/demo/web/gui/common/ProductDetailsPageBase.java
@@ -3,18 +3,28 @@ package com.qaprosoft.carina.demo.web.gui.common;
 import java.util.Date;
 import java.util.List;
 
-import com.qaprosoft.carina.demo.web.enums.FilterType;
 import org.openqa.selenium.WebDriver;
 
 import com.qaprosoft.carina.core.gui.AbstractPage;
+import com.qaprosoft.carina.demo.web.enums.FilterType;
 import com.qaprosoft.carina.demo.web.enums.ProductTabs;
 import com.qaprosoft.carina.demo.web.enums.SortDropdown;
+import com.qaprosoft.carina.demo.web.gui.components.Basket;
+import com.qaprosoft.carina.demo.web.gui.components.PhotoModal;
 
-public abstract class TableItemsPageBase extends AbstractPage {
+public abstract class ProductDetailsPageBase extends AbstractPage {
 
-    public TableItemsPageBase(WebDriver driver) {
+    public ProductDetailsPageBase(WebDriver driver) {
         super(driver);
     }
+
+    public abstract String getProductTitleText();
+
+    public abstract void clickOnBuyButton();
+
+    public abstract void moveToTitleText();
+
+    public abstract Basket getBasketMenu();
 
     public abstract void clickOnTab(ProductTabs product);
 
@@ -25,4 +35,8 @@ public abstract class TableItemsPageBase extends AbstractPage {
     public abstract List<Date> getListOfCommentsDate();
 
     public abstract boolean isOpinionsSortedByDate();
+
+    public abstract void clickOnPhotoTab(int index);
+
+    public abstract PhotoModal getPhotoModal();
 }
