+++ src/main/java/com/qaprosoft/carina/demo/web/gui/common/SearchPageBase.java
@@ -26,5 +26,10 @@ public abstract class SearchPageBase extends AbstractPage {
 
     public abstract void productNameFilterClick(FilterType filterType);
 
+    public abstract ProductListPageBase productTypeLinkClick(FilterType filterType);
+
     public abstract boolean checkSubcategoryTitleText(String subcategory);
+
+    public abstract boolean isTitleTextContainsProductType(FilterType filterType);
+
 }
