+++ src/test/java/com/qaprosoft/carina/demo/rozetka/RozetkaPDPTest.java
@@ -9,10 +9,11 @@ import com.qaprosoft.carina.demo.web.enums.Devices;
 import com.qaprosoft.carina.demo.web.enums.FilterType;
 import com.qaprosoft.carina.demo.web.enums.MenuCategory;
 import com.qaprosoft.carina.demo.web.enums.ProductTabs;
-import com.qaprosoft.carina.demo.web.gui.common.PhoneItemsPageBase;
-import com.qaprosoft.carina.demo.web.gui.common.PhonePageBase;
 import com.qaprosoft.carina.demo.web.gui.common.PhonesAndElectronicsPageBase;
+import com.qaprosoft.carina.demo.web.gui.common.ProductDetailsPageBase;
+import com.qaprosoft.carina.demo.web.gui.common.ProductListPageBase;
 import com.qaprosoft.carina.demo.web.gui.components.PhotoModal;
+import com.qaprosoft.carina.demo.web.gui.components.ProductFilter;
 import com.qaprosoft.carina.demo.web.gui.desktop.HomePage;
 
 public class RozetkaPDPTest implements IAbstractTest {
@@ -25,16 +26,17 @@ public class RozetkaPDPTest implements IAbstractTest {
         HomePage homePage = new HomePage(getDriver());
         homePage.open();
         homePage.clickOnClosePopupButton();
-        PhonesAndElectronicsPageBase phonesAndElectronicsPageBase = (PhonesAndElectronicsPageBase) homePage.clickOnCategoryMenu(MenuCategory.PHONES_TV_ELECTRONICS);
-        PhonePageBase phonePageBase = (PhonePageBase) phonesAndElectronicsPageBase.clickOnCategoriesLink(Devices.PHONES);
-        phonePageBase.selectBrand(FilterType.BRAND_APPLE);
-        PhoneItemsPageBase phoneItemsPageBase = phonePageBase.clickOnProductTitle(INDEX_ZERO);
-        phoneItemsPageBase.clickOnTab(ProductTabs.PHOTO);
-        phoneItemsPageBase.clickOnPhotoTab(INDEX_ZERO);
-        PhotoModal photoModal = phoneItemsPageBase.getPhotoModal();
+        PhonesAndElectronicsPageBase phonesAndElectronicsPageBase =
+                (PhonesAndElectronicsPageBase) homePage.clickOnCategoryMenu(MenuCategory.PHONES_TV_ELECTRONICS);
+        ProductListPageBase productListPage = phonesAndElectronicsPageBase.clickOnCategoriesLink(Devices.PHONES);
+        ProductFilter productFilter = productListPage.getFilter();
+        productFilter.selectFilter(FilterType.BRAND_APPLE);
+        ProductDetailsPageBase productDetailsPageBase = productListPage.clickOnDeviceTitle(INDEX_ZERO);
+        productDetailsPageBase.clickOnTab(ProductTabs.PHOTO);
+        productDetailsPageBase.clickOnPhotoTab(INDEX_ZERO);
+        PhotoModal photoModal = productDetailsPageBase.getPhotoModal();
         Assert.assertTrue(photoModal.isPhotoSliderPresent(INDEX_ZERO), "Photo modal isn't presented");
         photoModal.clickOnSlider();
         photoModal.clickOnModalCloseButton();
     }
-
 }
